- renaming namespace - changing path from blogController to be in the same pattern
